requirements.txt
pdfplumber>=0.10.0
python-docx>=1.1.0
google-genai>=0.1.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0

